Move code from init to base.py module.
Code insinide __init__.py indirectly imports jsonschema which adds a second or
so of load time to every script / module which directly or indirectly imports
that module.

This substantially reduced Python action wrapper run time.
